+++ lua/test/Lua/PrimaryTests.hs
@@ -13,7 +13,7 @@ import Foreign.C.String (withCString)
 import Foreign.Ptr (nullPtr)
 import Lua
 import Test.Tasty (TestTree, testGroup)
-import Test.Tasty.HUnit (Assertion, HasCallStack, testCase, (@=?) )
+import Test.Tasty.HUnit (Assertion, HasCallStack, assertBool, testCase, (@=?) )
 
 -- | Tests for unsafe methods.
 tests :: TestTree
@@ -25,6 +25,34 @@ tests = testGroup "Primary"
           lua_pcall l 0 1 0
           lua_type l (-1)
     ]
+
+  , testGroup "garbage-collection"
+    [ "stop, restart GC"  =:
+        -- first count should be larger
+        uncurry (>) `shouldHoldForResultOf` \l -> do
+          lua_createtable l 0 0
+          _  <- lua_gc l LUA_GCSTOP 0 0 0
+          lua_pop l 1
+          kb1 <- lua_gc l LUA_GCCOUNT 0 0 0
+          b1  <- lua_gc l LUA_GCCOUNTB 0 0 0
+          _   <- lua_gc l LUA_GCCOLLECT 0 0 0
+          kb2 <- lua_gc l LUA_GCCOUNT 0 0 0
+          b2  <- lua_gc l LUA_GCCOUNTB 0 0 0
+          return (b1 + 1024 * kb1, b2 + 1024 * kb2)
+    , "switch to generational GC"  =:
+        LUA_GCINC `shouldBeResultOf` \l -> do
+          lua_createtable l 0 0
+          GCCode <$> lua_gc l LUA_GCGEN 0 0 0
+    , "switch to generational and back to incremental GC"  =:
+        LUA_GCGEN `shouldBeResultOf` \l -> do
+          lua_createtable l 0 0
+          _ <- lua_gc l LUA_GCGEN 0 0 0
+          GCCode <$> lua_gc l LUA_GCINC 0 0 0
+    , "memory consumption should be between 0 and 10 kB" =:
+      (\count -> count > 0 && count < 10) `shouldHoldForResultOf` \l -> do
+          lua_gc l LUA_GCCOUNT 0 0 0
+    ]
+
   , testGroup "lua_stringtonumber"
     [ "converts a string to a number" =: do
         55 `shouldBeResultOf` \l -> do
@@ -48,3 +76,9 @@ shouldBeResultOf :: (HasCallStack, Eq a, Show a)
 shouldBeResultOf expected luaOp = do
   result <- withNewState luaOp
   expected @=? result
+
+shouldHoldForResultOf :: HasCallStack
+                      => (a -> Bool) -> (State -> IO a) -> Assertion
+shouldHoldForResultOf predicate luaOp = do
+  result <- withNewState luaOp
+  assertBool "predicate does not hold" (predicate result)